Add terminal preference order / customizability
The preference order is: argument, system-defined, urxvt, xterm.

If an argument is present, then it will be used. Otherwise, if $TERMINAL
is defined, which it is by some systems or could be by the user. If
i3-sensible-terminal is present, which has its own preference order,
then it will be used. If x-terminal-emulator is present,
which is the Debian/derived application that points to the default
terminal, that will be used. Then urxvt, then xterm.

This makes it so that there's a lower chance a terminal must be passed
in.