Add full list of programs from PATH
dmenu (or at least suckless-tools's dmenu) actually comes with the
application to list all executables on the path. It's called
dmenu_path. If this is present (which it presumably is), it is used
instead. If it is not, then the fallback will be /usr/bin. /usr/bin is a
superset of /bin, so it will include more.