Add automatic graphical v. terminal detection
ldd is searched for libX11 given the application name. If it is there,
then it is opened directly. If it is not, then it is opened with
terminal.

With this behaviour, the original purpose of ~/.dmenu-terminal is
deprecated. However, this method of detection can occasionally get it
wrong, as in the case of `vim` when `gvim` is installed, and `firefox`.
For these cases, ~/.dmenu-terminal is re-purposed as an *override* file.
That way it can be used as little as possible. I will update the
documentation in a future commit, but it's simple:

application [0 for console, 1 for graphical]
application2 1
app3 0